Use printf modifiers %z and %t for size_t and ptrdiff_t.
They were added in C99, so should be available in C++11.

For int64_t / uint64_t use PRId64 / PRIu64 macros from <cinttypes>.

Also fix some formatting issues.

+++ src/debug/dump_adfa.cc
@@ -52,7 +52,7 @@ void dump_adfa(const DFA& dfa) {
         }
         fprintf(stderr, "  n%u [height=0.2 width=0.2 label=\"", st2idx[s]);
         if (s->fill && action != Action::Kind::MOVE) {
-            fprintf(stderr, "F(%u) ", (uint32_t)s->fill);
+            fprintf(stderr, "F(%zu) ", s->fill);
         }
         if (action == Action::Kind::RULE) {
             const Rule& r = dfa.rules[s->action.info.rule];
